requirements.txt
Flask
python-dotenv
openai
requests
beautifulsoup4
langchain
langchain-openai
langchain-community